EES-2911 Move release amendment logic into extension methods
